# gridclan-backend/src/main/java/com/gridclan/dto/AuthResponse.java
package com.gridclan.dto;

import lombok.*;
import java.util.UUID;

@Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
public class AuthResponse {
    private String accessToken;
    private String refreshToken;
    private String role;
    private UUID userId;
}